+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Release history
 ---------------
 0.6.x (unreleased)
 ++++++++++++++++++
-- improved: dnswanip error reporting
+- improved: dnswanip error reporting now includes dns information
+- improved: fix for bug `#144 <https://github.com/infothrill/python-dyndnsc/issues/144>`_
 
 0.6.0 (February 21st 2021)
 ++++++++++++++++++++++++++